requirements.txt
Django==3.0.7
djangorestframework==3.11.0
Markdown==3.2.1
django-filter==2.2.0
djangorestframework-simplejwt==4.4.0
pytest-django==3.9.0
python-decouple==3.3
boto3==1.14.12
Pillow==7.1.2
django-storages==1.9.1